Fixes mainly due to sqlite migration

+++ RunnersPal.Core/Models/RunLog.cs
@@ -13,7 +13,7 @@ namespace RunnersPal.Core.Models
         public static dynamic Route(this object runLog)
         {
             dynamic dynRunLog = runLog as dynamic;
-            return new Route().Single(dynRunLog.RouteId);
+            return new Route().Single((long)dynRunLog.RouteId);
         }
     }
 }
